requirements.txt
google-api-python-client>=2.130
google-auth>=2.35
google-auth-oauthlib>=1.2.0
python-dateutil>=2.9
typer>=0.12
rich>=13
python-dotenv>=1.0.1
google-generativeai>=0.7.2

# ML and data analysis dependencies
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
pandas
pyarrow
fastparquet

# Web interface
streamlit>=1.28.0
plotly>=5.15.0
